requirements.txt
requests~=2.26.0
yfinance==0.1.70
PyYAML~=6.0